requirements.txt
Django==1.7.2
MySQL-python==1.2.5
Unidecode==0.04.17
bibliograph.core==1.0.3
bibliograph.parsing==1.0.1
bibtexparser==0.6.0
django-autocomplete-light==1.4.14
django-crispy-forms==1.4.0
gnureadline==6.3.3
ipython==2.3.1
kitchen==1.1.1
lxml==3.4.1
requests==2.5.1
wsgiref==0.1.2
